requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
feedparser
jinja2
pydantic
python-dotenv
PyYAML
httpx
